refactor: Lp 댓글 관련 기능 LpComments로 분리
